Convert to CMake
Prior to this commit, Make was used for the firmware but CMake was
desired for its simplicity. This commit modifies each module to build
with CMake. The code has also been reorganized into a "platform/module"
structure (rather than a "module/platform" structure) for ease of building
all firmware for an entire platform when, in the future, cars other than
the KIA Soul have their own firmware.